Moved editorDimensions to ConnectionStage
