+++ src/types/ConnectionStageTypes.ts
@@ -1,12 +1,11 @@
 import { MouseEvent } from "react";
 import { Connection, ConnectionPosTable } from "./NodeEditorTypes";
-import { clientDimensions, DragOffset } from "./utilTypes";
+import { DragOffset } from "./utilTypes";
 
 export interface ConnectionStageProps {
   zoom: number;
   setZoom: (zoom: number) => void;
   setConnections: (cons: Connection[]) => void;
-  editorDimensions: clientDimensions;
   nodeEditorOffset: { x: number; y: number };
   updateMousePath: (e: MouseEvent) => void;
   panningOffset: DragOffset;
